requirements.txt
openai==1.54.4
gspread
python-dotenv
crawl4ai